+++ packages/server/src/Hocuspocus.ts
@@ -1,5 +1,5 @@
 import * as decoding from 'lib0/decoding'
-import WebSocket, { WebSocketServer } from 'ws'
+import WebSocket, { AddressInfo, WebSocketServer } from 'ws'
 import { createServer, IncomingMessage, Server as HTTPServer } from 'http'
 import { Doc, encodeStateAsUpdate, applyUpdate } from 'yjs'
 import { URLSearchParams } from 'url'
@@ -55,7 +55,7 @@ export class Hocuspocus {
     onDestroy: () => new Promise(r => r(null)),
   }
 
-  documents = new Map()
+  documents: Map<string, Document> = new Map()
 
   httpServer?: HTTPServer
 
@@ -137,7 +137,7 @@ export class Hocuspocus {
   async listen(
     portOrCallback: number | ((data: onListenPayload) => Promise<any>) | null = null,
     callback: any = null,
-  ): Promise<void> {
+  ): Promise<Hocuspocus> {
     if (typeof portOrCallback === 'number') {
       this.configuration.port = portOrCallback
     }
@@ -205,17 +205,37 @@ export class Hocuspocus {
     this.httpServer = server
     this.webSocketServer = webSocketServer
 
-    await new Promise((resolve: Function, reject: Function) => {
+    return new Promise((resolve: Function, reject: Function) => {
       server.listen(this.configuration.port, () => {
         if (!this.configuration.quiet && process.env.NODE_ENV !== 'testing') {
           this.showStartScreen()
         }
 
-        this.hooks('onListen', { port: this.configuration.port })
-          .then(() => resolve())
+        this.hooks('onListen', { port: this.address.port })
+          .then(() => resolve(this))
           .catch(error => reject(error))
       })
     })
+  }
+
+  get address(): AddressInfo {
+    return (this.httpServer?.address() || {
+      port: this.configuration.port,
+      address: '127.0.0.1',
+      family: 'IPv4',
+    }) as AddressInfo
+  }
+
+  get URL(): string {
+    return `127.0.0.1:${this.address.port}`
+  }
+
+  get webSocketURL(): string {
+    return `ws://${this.URL}`
+  }
+
+  get httpURL(): string {
+    return `http://${this.URL}`
   }
 
   private showStartScreen() {
@@ -224,8 +244,8 @@ export class Hocuspocus {
     console.log()
     console.log(`  ${kleur.cyan(`Hocuspocus v${meta.version}${name}`)}${kleur.green(' running at:')}`)
     console.log()
-    console.log(`  > HTTP: ${kleur.cyan(`http://127.0.0.1:${this.configuration.port}`)}`)
-    console.log(`  > WebSocket: ws://127.0.0.1:${this.configuration.port}`)
+    console.log(`  > HTTP: ${kleur.cyan(`${this.httpURL}`)}`)
+    console.log(`  > WebSocket: ${this.webSocketURL}`)
 
     const extensions = this.configuration?.extensions.map(extension => {
       return extension.constructor?.name
@@ -280,7 +300,7 @@ export class Hocuspocus {
         return
       }
 
-      document.connections.forEach(({ connection } = { connection: Connection }) => {
+      document.connections.forEach(({ connection }) => {
         connection.close(ResetConnection)
       })
     })
@@ -536,7 +556,10 @@ export class Hocuspocus {
   private async createDocument(documentName: string, request: IncomingMessage, socketId: string, connection: ConnectionConfiguration, context?: any): Promise<Document> {
     if (this.documents.has(documentName)) {
       const document = this.documents.get(documentName)
-      return document
+
+      if (document) {
+        return document
+      }
     }
 
     const document = new Document(documentName)